Handle temperatures as integers
representing thousandths of a degree.  Avoids having to deal with
floats, and all the oddities they bring.

+++ webserver/dummy_sensor_test.go
@@ -2,16 +2,16 @@ package webserver_test
 
 // dummy sensor implementation for testing webserver
 type dummySensor struct {
-	temp float64
+	temp int
 }
 
-func (s *dummySensor) Temperature() float64 {
+func (s *dummySensor) Temperature() int {
 	return s.temp
 }
 
 func (s *dummySensor) Close() {
 }
 
-func (s *dummySensor) SetTemperature(value float64) {
+func (s *dummySensor) SetTemperature(value int) {
 	s.temp = value
 }
